Simultaneously dial a list of contacts associated with the organization
Summary:
Providers part of the organization manage a list of phone numbers associated with the org. This list of phone numbers is what is simultaneously dialed out when a patient calls the organization phone number.

Note that there is a current limitation to the simultaneous dialing, where if any one of the phone numbers dialed out goes to voicemail, it is considered "answered" and hangs up the other calls that are attempting to be made. We however consider this an OK limitation for now given that we will learn how people are using the product to do anything more sophisticated.

Another thing to note is that we are managing a list of "contacts" to call not a list of "providers" to call. This is by design so as to ensure that you can add a landline to dial out when a patient calls in.

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D338