Reserving a proxy phone number from a pool of available phone numbers for outgoing calls
Summary:
This change introduces the ability to reserve a proxy phone number from a pool of available numbers to make an outgoing call. The number is reserved for a specified period of time and then considered available after the (expiration time + grace period). The grace period is there to account for the provider starting the call close to the end of the expiration time.

Currently we reserve a single phone number for an outgoing call to an external entity. Once reserved, anyone can call that number; the call does not have to originate from the phone of the caller. This is because, especially on the web, it is not necessary that the call originates from a number associated with the provider's account.

In the future we could enforce stricter controls by requiring the provider to confirm the number from which they will be placing the call, or even being required to enter a pin to ensure that there are no collisions and the person attempting to call the patient is the one that places the call.

You can read more about the reasoning behind the current flow in this document:
https://docs.google.com/document/d/1NG_NOUM9YaVEmtq8PB9vwzX_9tvARYv0bR7lVB_R94Q/edit

I've also introduces a phone lib which for now does very basic validation and formats only US numbers. Over time, we can build this library to include transformations to/from friendlier versions that users will be inputting (like (206) 877 3590 -> +12068773590)

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D339