Uploading Twilio media objects to S3
Summary:
Asynchronously uploading recordings and SMS media attachments from the worker  rather than at the time of receiving the recording/media attachments because we can. With sendgrid, the email attachments came in as data part of the request rather than URLs with the attachments already uploaded to sendgrid's servers.

The work still remaining here is to pass the client the actual URL with data of the appropriate size, as well as to delete the data from twilio.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D358