Enable setting of logger from the caller for creating a service to manage consul locks
Summary:
Our restapi boxes in production had reached a state where they were unable to acquire a consul lock since ~Dec 14 2015. We were not notified of this problem for a couple reasons.

1. While an error was being logged non-stop, the sns handler for notifying of error logs was not setup for the consul service's logger.
2. We are not currently logging on the lack of metrics for the workers that rely on the consul lock, but instead logging on the transition of crossing a particular threshold and persisting the state of being above threshold.

This fix is to pass through a logger at the time of instantiating the handler so that the sns error logging handler gets used in the consul logger.

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D365