baymaxgraphql: fix authenticate mutation
Summary: The problem was that updating ResolveParams.Context didn't mean that all further queries saw the new context. The "fix" (read kludge) was to always have a valid account struct in the context and copy over it rather than updating the context. Not sure of a better way at the moment.. but should revisit as the GraphQL pkg matures.

Reviewers: kunaljham, michaelraines

Reviewed By: kunaljham, michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D405