Ensure to check all possibile entities for routing external messages to threads
Summary:
When a thread is deleted today, we only mark the thread as deleted, not the entity belonging to the organization. This bug is tracked here: https://app.asana.com/0/12387442313872/90496453203782

Consequently, this exposed a bug in the routing layer where we were not handling the possibility that an inbound message from an external entity could be routed to multiple threads. This is possible because a user may start one thread, and then add a contact that is already part of another thread.

This fix addresses that and also solves the bug where when a thread is deleted, if we start a new thread for the same contact, it would create a new entity, but the routing layer would take the first entity that matched the contact value rather than consider all of them, and the first entity was actually the deleted entity.

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D453