Magic threads
Summary:
- onboarding thread
- support thread
- initial internal thread

Onboarding thread works by having a new table that tracks which threads are onboarding. The invite and excomms services publish to SNS topics events that are relevant to the onboarding. There's a new worker in the threading service that watches for those events as well as published messages from the threading service to see if steps are completed or skipped.

The support thread is created as two threads in separate orgs. There's a new table that provides linking between them. If the threading service sees a post to a thread that has a linked pair then it'll mirror the message in both threads swapping out the primary entity as the sender in the remote thread.

The internal thread is a normal thread with the organization itself as the primary entity. Changes were made to not return contacts as endpoints for organizations on threads.

Reviewers: michaelraines, kunaljham

Reviewed By: kunaljham

Differential Revision: https://phabricator.corp.sprucehealth.com/D491