Worker to block accounts published to topic
Summary:
In an effort to clean up the spruce support org threads, as well as make blocking of an account easier, created a worker that takes the necessary steps to block an account. Created a separate service for such operational work. If this makes sense, will also move the resource cleaner in the excomms layer over here. Additionally plan to add a worker to deprovision phone numbers in non-prod environments after every 24 hours so that we are not paying for a large number of phone numbers.

While the blocking of an account has been made simpler via the worker here, the act of "identifying" accounts to block is still a manual process. I plan to add a script now that makes it easy to publish a list of email addresses to block. Over time the hope is that we can also automate the identifying suspect accounts which we can move into the "suspended" state, verify they are indeed suspect, and then confirm them to be blocked.

Reviewers: michaelraines, samuel

Reviewed By: michaelraines, samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D549