Introduce configurable setting for whether or not to reveal sender
Summary: Griffin Concierge Medical requested this feature. Given the relative easy with which we can introduce settings, thought we'd get this into the product such that we can reveal the sender for organizations that want this setting when communicating over external communications. This also makes the behavior consistent across email and SMS where currently over Email we were revealing the sender but not over SMS.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D556