Change in send all calls to voicemail behavior
Summary:
The setting will now exist for the phone number rather than globally for the provider in the organization. Previously it was serving the purpose of a DND mode where if the provider did not want to receive calls they'd turn that switch on. However, while that was the intention the toggle is attempting to be used as a way to divert all calls to the spruce phone number to voicemail.

So migrating this setting to actually work at the phone number level - where if the toggle is turned on for the phone number in the org, all calls will be directed to voicemail.

There is a migration script that is intended to be run once all new clients are out.

Reviewers: michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D582