Script to ingest all Layouts into the system
Summary: Until we have the web component to manage the Layouts in the system, starting with this script to have a way to ingest the category set into the system. We can incrementally add support for updating single SAML at a time if we want. But right now the way this script works is that you give it the zipped content directory and the categories you want visits to belong to and it updates it in the particular environment.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D694