erx/dosespot: remove dependency on restapi models
Summary:
Created a new lib/dosespot that's the low-level DoseSpot client,
and stripped out and moved the old lib/erx to cmd/svc/restapi/erx so
that it's clear it's restapi specific.

I wanted to avoid having to touch any of the restapi code so having a
lib with the same interface was the simplest and felt least likely to
cause bugs.

Reviewers: michaelraines, kunaljham

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D727