Only include entity information for entity if queries for that particular node
Summary:
okay I think I found the issue with why contacts are sometimes populated and sometimes not on the backend. When we request for addressableEntities for an external thread, we are requesting for members of the org which returns ​_all_​ members (internal and external). The request we make is the entity information to include MEMBERS and CONTACTS at Depth = 0. This means that only the top level entities (i.e, the org) should contain the MEMBERS and CONTACTS while the members should not contain MEMBERS and CONTACTS. While it is true that the members (internal and external nodes) don’t contain the contact information, the `IncludedInformation` field of the members indicates that CONTACTS are returned, and all entities returned are cached at the baymaxgraphQL layer.

So when we then go to request the primaryEntity at the thread level, we find the Entity in the cache, and the `IncludedInformation` field indicates that CONTACTS are populated when they are not, so we consider that a cache hit and return the entity but without the actual contact information populated.

Tested this in staging to find it to be true:
```
ubuntu@ip-10-1-23-115:~$ ./entitycachetest
Entities:
-- Afterhours testing [CONTACTS MEMBERS]
---- CONTACTS: [&Contact{ContactType:PHONE,Value:+12065391901,Provisioned:true,ID:entityContact_0DVD0BKJG3000,Label:,}]
-- MEMBERS:
---- Kunal Jham, CN [CONTACTS MEMBERS]
---- CONTACTS: []
---- Spruce Support [CONTACTS MEMBERS]
---- CONTACTS: []
---- Setup [CONTACTS MEMBERS]
---- CONTACTS: []
---- ANONYMOUS [CONTACTS MEMBERS]
---- CONTACTS: []
---- (617) 515-1344 [CONTACTS MEMBERS]
---- CONTACTS: []
---- Nadine Haija, DMD [CONTACTS MEMBERS]
---- CONTACTS: []
---- (206) 877-3590 [CONTACTS MEMBERS]
---- CONTACTS: []
---- Setup Assistant [CONTACTS MEMBERS]
---- CONTACTS: []
---- RESTRICTED [CONTACTS MEMBERS]
---- CONTACTS: []
---- Patient Jham [CONTACTS MEMBERS]
---- CONTACTS: []
---- Spruce Number [CONTACTS MEMBERS]
---- CONTACTS: []
```

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D735