RAM: Refactor querying of entities
Summary:
Fundamentally, the requirement is for the caller to specify the `directory.LookupEntitiesRequest` or `directory.LookupEntitiesByContactsRequest` object to give the caller the most flexibility to query for entities as they see fit.

Consequently, the ResourceAccessor interface now has the following methods to query for entities:
```
	Entities(ctx context.Context, req *directory.LookupEntitiesRequest, opts ...EntityQueryOption) ([]*directory.Entity, error)
	EntitiesByContact(ctx context.Context, req *directory.LookupEntitiesByContactRequest) ([]*directory.Entity, error)
```

versus what we previously had:

```
	ActiveEntity(ctx context.Context, entityID string, entityInfo []directory.EntityInformation, depth int64) (*directory.Entity, error)
	Entity(ctx context.Context, entityID string, entityInfo []directory.EntityInformation, depth int64) (*directory.Entity, error)
	Entities(ctx context.Context, orgID string, entityIDs []string, entityInfo []directory.EntityInformation) ([]*directory.Entity, error)
	EntitiesByContact(ctx context.Context, contactValue string, entityInfo []directory.EntityInformation, depth int64, statuses []directory.EntityStatus) ([]*directory.Entity, error)
	PatientEntity(ctx context.Context, a *models.PatientAccount) (*directory.Entity, error)
	UnauthorizedEntity(ctx context.Context, entityID string, entityInfo []directory.EntityInformation, depth int64) (*directory.Entity, error)
```

The motivation for this change was to take advantage of the type filtering on root and child types that would improve overall performance for users (we no longer query for _all_ entities when trying to get members of org) and fit that in a way that did not require extending the existing method set or modifying method signatures.

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D737