Only allow spruce visit attachments for a patient on iOS.
Summary: Tracking login for a user on each platform so that we know whether a patient that signed up did so on android, ios, etc. Its also useful to have this information for the future. This enables us to only allow visit attachments for a patient once an account has been created and if the patient is on iOS

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D757