Media service integration for visits
Summary: The visits now contain real photo and async video urls when returning answers to questions in a visit or the visit preview for the provider. Currently the calls are being made directly to the media service with the assmption that media service will be responsible for authorization checks. If we decide to have a helper library that sits atop the media service then the calls to the media service would have to be updated to fit the pattern.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D768