Stub to mark all threads as read
Summary:
Introducing a stub to mark all threads as read so that client side can integrate this feature end to end to ensure that it works.
Also introducing a flag to control whether or not the act of shaking your phone to mark threads as read is allowed at an org level.
Making it an org level setting so that we can control this for each org

Reviewers: samuel, michaelraines

Reviewed By: samuel, michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D773