Query for entity without filtering out inactive users
Summary: If you have a member of the team that has been removed, and that member had sent a message on a thread, then it is possible that the thread did not load for other members of the team. Specifically because we were querying for the actor of a message expecting it to be ACTIVE when not all actors may be active.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D797