Add setting for video calls
Summary: Started with the idea that the video call setting would be done directly through the settings client so created a caching version, but that seemed way too invasive and messy so created the idea of feature flags in the context that can be lazily evaluated. It's not great, but felt cleaner.

Reviewers: michaelraines, kunaljham

Reviewed By: kunaljham

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D821