Optimistically allow spruce visit attachments when patient has not created account
Summary:
The situation where the patient has not created the account and the provider
tries to send the patient a spruce visit attchment is probably low. But the situation
where a customer is trying to demo spruce visits and cannot get the spruce visit option
to show up until the patient app is created on iOS is currently high.
So optimize for spruce visit showing up as attachment when + button tapped until we know
for certain that patient cannot have spruce visit attached because they are on android

Reviewers: michaelraines, samuel

Reviewed By: samuel

Differential Revision: https://phabricator.corp.sprucehealth.com/D816