Unifying custom voicemail greeting flow
Summary:
Now that iOS enables the provider to set a custom voicemail, we want to make sure that the provider is able to query any custom greetings we have already uploaded. However that is currently not possible as the free text response of the setting contains a media URL rather than just the mediaID which is what the client is expecting. Secondly, the client is also setting the mediaID in the free text response and not the mediaURL  which is the correct value to set.

Unfortunately up until this point we have been setting the media URL when manually uploading the custom voicemail and the excomms layer also expects the mediaURL to exist.

So this change does a few things:
1. Excomms layer now expects mediaID rather than media URL
2. Custom voicemail greeting is driven off single setting config rather than separate ones for after hours and standard voicemail configuration. This enables provider to set custom voicemail in app and tell us that they want the afterhours configuration.
3. Migration script to migrate all afterhours and standard voicemail settings to contain mediaID and to all be stored under the `voicemail_option` setting.

Reviewers: janky, michaelraines, samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D839