Intake library updates for Baymax
Summary:
This PR updates the cross platform library to work for Spruce visits in baymax. Note that it is *not* backwards compatible to work with Spruce derm. If we ever have to make changes to the intake library for Spruce Derm we are going to have to checkout the tag https://github.com/SpruceHealth/backend/releases/tag/derm_intake_lib and work off of this tag.

All tests have been updated and the coverage did not change even though we added functionality and took some away
```
➜  manager git:(baymax_intakelib) ✗ go test -cover
PASS
coverage: 76.3% of statements
ok  	github.com/sprucehealth/backend/libs/intakelib/manager	0.300s
```

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D833