Support updating of first and last name for provider profile
Summary:
Only allowing the provider to update their display name introduced the problem of us not knowing how to update their initials when they do not have an image because the initials are constructed at read time from their first and last name. Consequently, in order to give the provider complete control we are now letting them update their first and last name in addition to their display name like so:
https://www.dropbox.com/s/hbo26rina97z6aj/Screenshot%202016-07-11%2017.34.28.png?dl=0

This change allows the client to do that via the mutation.

Reviewers: michaelraines, samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D846