Command to make it easy to enable the packages for our customers
Summary: Wanted to make it easy for us to enable the two packages that we are selling today - answering service and digital care solution. Over time it may make sense to bundle the settings into these two packages such that with each feature addition we dont have to retroactively enable the feature for anyone that falls into one of the tiers. But until we figure out what the right package is and enable settings via admin panel, this will at the very least provide a fool proof way of turning on a package for an org

Reviewers: michaelraines, samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D847