Increase read timeout for media uploads
Summary:
Android is not compressing videos before uploading them which is leading to larger video sizes more frequently than iOS. This change allows a video upload to continue for 5 minutes to increase the likelihood that video uploads complete. Note that video sizes on android are limited to 100MB.

This means that on LTE (which typically have between 2-5 Mbps of upload speed), it should take anywhere between 2 mins 40 seconds and 6 minutes 40 seconds to upload. This is definitely not ideal, and we will be looking into client side compression of videos before upload in 1.6.1.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Differential Revision: https://phabricator.corp.sprucehealth.com/D868