Adding a webhook for twilio to interact with when it encounters errors in any communication with our servers
Summary:
Adding a simple API endpoint that twilio can hit to inform us of errors that twilio has interacting with our API for calls and SMS.

Twilio has a mechanism to trigger a webhook upon a configured alert trigger. So the idea is that twilio would hit this endpoint on errors and then we can use librato to alert us of issues. This will help inform us of any "an application error has occured" issue on patient or provider side, as well as any custom voicemail issues, etc.

I know we want to improve our metric publishing mechanism, but I want to get this change out there ASAP given that providers have been complaining of call related issues in the past week and this will help inform us when there are issues we should care about.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D882