Returning back valid response in the event of an invalid from phone number for an incoming call.
Summary: Currently an invalid phone number (like an international phone number) for incoming calls gets treated as a server error on Twilio's side because it is unable to retrieve the instructions for how to handle such an incoming call. Returning back valid TwIML in this situation will help reduce most of the errors that we are seeing with twilio and help us make our alarm on the error count sensitive so that we catch all customer facing issues and actually alarm on true customer facing issues.

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Differential Revision: https://phabricator.corp.sprucehealth.com/D883