gRPC based health checks
Summary:
Using the gRPC server itself for health checks gives us a more reliable
way to check on the status of the service (instead of just seeing of the
port is listening for connections).

The grpchealthcheck cli will be on svc cluster instances and can be
called by consul to check on the status of a service as configured in
registrator.

Reviewers: michaelraines, kunaljham

Reviewed By: kunaljham

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D886