Incoming call improvements
Summary:
The reason for the updated 0 rows error we were seeing in production is because we were trying to mark a call as completed while processing voicemail and also in the status of an inbound call. Both requests were being made in the same second, so in one of the updates no rows were getting updated since the data with which the row was trying to get updated was the same as what was in the database.

This change makes it the sole responsibility of the inbound call status processing request to mark a call as completed.

Also tracking when a voicemail is left so that we can use that information to our benefit in the future.

Reviewers: samuel, michaelraines

Reviewed By: samuel, michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D896