Core EMR component
Summary:
The core EMR component is responsible for creating standard or secure threads for the organization based on the incoming ADD sync event. Sync events are processed by a worker which creates the threads and will also be responsible for either publishing a notification for the created patient or writing the patient link to the external source itself.

Things to note:
- Only `ADD` sync event support which enables adding of multiple patients.
- `ADD` event is idempotent
- Any interaction with the EMR will layer on top of this. The EMR interaction layer will be responsible for publishing a message that contains the messages to add.
- Only supporting secure threads creation with "deferred" invite where the invite is not sent at the time of secure thread creation but when the provider presses the "Resend invite" button at the thread level

Reviewers: samuel, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D913