Spruce-Hint Integration
Summary:
This change makes it possible to:
1. Initiate a sync with a practice's hint account such that secure/standard threads are created for all existing patients in a practice's account
2. Sync new patients that are created after the initial sync into the practice account

The `InitiateSync`is intentionally a service method so that we can decouple the act of sync configuration (service method not implemented yet) from when the sync process start. Additionally, I intentionally send a message to an SQS queue to initiate the sync knowing that the sync can take a while to complete.

At this point, I'm ready to test this interaction with hint.

What's next:
1. Deploying the `patientsync` service to dev and getting this working with a test practice account
2. Figuring out an automated O-Auth flow in integration with the web signup flow so that we can start the "connect" flow from the Hint side for a practice to connect a new or existing stripe account with Hint.

Reviewers: michaelraines, samuel

Reviewed By: michaelraines, samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D928