Track deviceID and platform for auth tokens that are created
Summary:
Previous to this PR it was possible for a user that is logging in on multiple devices to get the token of another device. Here's how:
1. Device1 makes a request with Token1 that is outside of the token refresh window so the token is rotated and a shadow token is created for Token1.
2. Device1 gets the rotated token Token1' in the request, but also has other in-flight requests with Token1.
3. When Device1 makes requests with the shadow token, the server attempts to return the latest created active token to the user.
4. If the user logged in to Device2 after having logged in to Device1, then the token of Device2 would be returned to Device1.  This happens because we don't track deviceID to return to the user the active token created for the particular device the user is making the request from.
5. If the user goes back to Device2, the token on Device2 çan also be rotated given that it is out of the token refresh window, resulting in the Token that Device1 has becoming a 24 hour expiring token given that it is now a shadow token.
6. If the user now goes back to Device1 after ~24 hours, the token would have expired.

What seems like an edge case for the number of steps required to get to it is actually a very normal use case for users using Spruce on iOS and web. There are currently 288 such users. And what they can experience is:
1. A premature logout from iOS even when they have secure TouchID on (if the SHORT token from web gets transferred to iOS)
2. A 1 year valid token on the web (if the LONG token from iOS gets transferred to Web)

We fix this by tracking deviceID and platform in the auth token table and ensuring only to return an active token created for the deviceID the user is making the request from.

Reviewers: michaelraines, samuel

Reviewed By: samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D947