Allow deleting of a secure conversation where the patient has not yet created an account
Summary: The most common reason for a provider to want to delete a secure conversation is beacuse they sent the invite to the wrong patient or the patient is not going to create an account. This change makes it possible to delete a secure conversation only if the patient has not yet created an account.

Reviewers: michaelraines, samuel

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D914