Materialized saved query indexes
Summary:
This change adds individual indexes for every saved query. This means
that on all operations that mutate a thread we must update all possible
saved queries. This generally involves pulling all the saved queries for
all members (which means following membership relationships as an org
can be a member of a thread) and seeing which ones match.

I decided to use the same MySQL database as the primary store (as
opposed to DynamoDB) as it gives us guarnatees around updating
normalized total and unread counts for the saved queries. It's still
possible for the saved query index to not totally reflect reality, but
there's checks in place when querying to detect such situations.

Right now all updates to the indexes are synchronous with any request
that mutates the thread. This is so that the user sees an updated world
view right away. However, we'll likely want to be smarter about this in
the future but only synchronously updating the saved queries for the
person that initiated the operation and asynchronously for everyone else
(as they won't see a difference).

the riskiest part is a thread showing up in a saved query index even
after a member is removed. I've tried to mitigate this by verifying
membership on read as well. It's possible I missed some places though so
that's something to pay attention to. The person won't be able to
actually access the thread if it's wrong, but they may see it in the
list when they shouldn't.

Reviewers: michaelraines, kunaljham

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D955