Accurate badge count (#1093)
* Accurate badge count

To keep track of all threads in the set of notification saved queries,
this change adds a new saved query that acts as a catch all bucket for
them. It's treated differently in that the query isn't used, but rather
any place we add a thread to a saved query that's set to notify we also
add it to the catch all one.

Another new feature of saved queries is a hidden flag that is used to
filter what saved queries are shown to the user in the inbox view. The
ones that are hidden can still be used to set notification preferences.

* Fix panic

* Add rebuildsavedqueries command to baymaxadmin cli

* Fix typo

* Add badge count back to notifications

* Fix notification tests

* Obey badge_count setting