Remove the idea of "ID" aka "URL" from the storage lib. (#1187)
There's always existed the confusion of the ID (sometimes called URL)
returned by the storage Put. It was designed to be able to have a
"handle" to an object rather than an ID so that if the media needed to
be moved the actuall location existed with the handle. This caused too
much confusion however especially when called URL (it was only a URL in
the case of the S3 store).

In order to avoid this confusion and to simplify the interface the
storage pkg now only deals with IDs. The Put methods still return a
string but not it's the ID. This is for compatibility. The Get methods
support both a plain ID and the old ID/URL handle and fall back.

The threading service translates old style IDs (the media service was
storing the ID while excomms was storing the handle URL) to the plain
ID.