admin: materialize saved query templates (#1259)
Organizations start out with no saved query templates and a default set
is used. They don't exist, however. This means it's not possible to
update/delete them. This change adds an admin mutation to materialize
the default set as concrete for an organization. After this it's
possible to modify as needed.