threading: fix unread notification count if we know it's wrong (#1304)
When fetching all the saved queries for an entity we can get the min and
max possible counts for the notification unread saved query. If the
actual value is out this range then trigger a rebuild of that saved
query.

There's no way to know for sure if the count is exactly right due to
duplication of threads across saved queries, but this should fix the
most annoying case of the badge count not going to zero.